Fix bnc#588910
Running lilo in an LVM configuration on top of dm-multipath will
fail, because lilo attempts to convert the PreP boot partition to
a FAT partition, then mkfs.msdos the partition. However, since
dm-multipath is using that partition, the mkfs.msdos fails. Fix
this by checking for a parent dm device is the mkfs.msdos fails
and using the parent dm device instead of the single path.

Signed-off-by: Brian King <brking@linux.vnet.ibm.com>